+++ lldb/CMakeLists.txt
@@ -61,6 +61,10 @@ if (LLDB_ENABLE_PYTHON)
     "Path to python interpreter exectuable, relative to python's install prefix")
   set(cachestring_LLDB_PYTHON_EXT_SUFFIX
     "Filename extension for native code python modules")
+  set(cachestring_LLDB_PYTHON_DLL_RELATIVE_PATH
+    "Relative path from LLDB executable to Python DLL directory (Windows only)")
+  set(cachestring_LLDB_PYTHON_RUNTIME_LIBRARY_FILENAME
+    "Filename of Python runtime library to search for, e.g. python311.dll (Windows only)")
 
   if (LLDB_ENABLE_PYTHON_LIMITED_API)
     set(stable_abi "--stable-abi")
